# DaJaMa Games/war/api/DaJaMaSwagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "DaJaMa Api",
        "description": "API que ofrece una manera de gestionar tus juegos favoritos",
        "version": "1.0.0"
    },
    "host": "1-dot-apidajama.appspot.com",
    "schemes": [
        "http"
    ],
    "basePath": "/api",
    "produces": [
        "application/json"
    ],
    "paths": {
        "/GameList": {
            "get": {
                "summary": "Listas de juegos",
                "description": "Devuelve todas las listas de juegos\n",
                "responses": {
                    "200": {
                        "description": "Un array de lista de juegos",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/gameList"
                            }
                        }
                    }
                }
            },
            "post": {
                "summary": "Crear lista de juego",
                "description": "Crea una lista de juego para un usuario\n",
                "responses": {
                    "201": {
                        "description": "Lista de juegos creada",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "444": {
                        "description": "Nombre de la lista ya existente"
                    }
                }
            }
        },
        "/GameList/{gameListName}": {
            "get": {
                "summary": "Lista de juegos específica",
                "description": "Devuelve una lista de juegos deseada\n",
                "parameters": [
                    {
                        "name": "gameListName",
                        "in": "path",
                        "description": "Nombre de la lista a obtener información",
                        "required": true,
                        "type": "string",
                        "format": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Una lista de juegos",
                        "schema": {
                            "$ref": "#/definitions/gameList"
                        }
                    },
                    "404": {
                        "description": "Lista no encontrada"
                    }
                }
            },
            "put": {
                "summary": "Modificar una lista",
                "description": "Modifica una lista de juegos existente\n",
                "parameters": [
                    {
                        "name": "gameListName",
                        "in": "path",
                        "description": "Nombre de la lista que desea modificar",
                        "required": true,
                        "type": "string",
                        "format": "string"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Lista modificada"
                    },
                    "404": {
                        "description": "Lista no encontrada"
                    }
                }
            },
            "delete": {
                "summary": "Eliminar una lista",
                "description": "Elimina una lista de juegos existente\n",
                "parameters": [
                    {
                        "name": "gameListName",
                        "in": "path",
                        "description": "Nombre de la lista que desea eliminar",
                        "required": true,
                        "type": "string",
                        "format": "string"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Lista eliminada"
                    },
                    "404": {
                        "description": "Lista no encontrada"
                    }
                }
            }
        },
        "/GameList/{gameListName}/games": {
            "get": {
                "summary": "Juegos de una lista",
                "description": "Devuelve todas los juegos de una lista\n",
                "parameters": [
                    {
                        "name": "gameListName",
                        "in": "path",
                        "description": "Nombre de la lista a obtener información",
                        "required": true,
                        "type": "string",
                        "format": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de juegos",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/game"
                            }
                        }
                    },
                    "404": {
                        "description": "Lista no encontrada"
                    }
                }
            },
            "post": {
                "summary": "Añade nuevo juego",
                "description": "Añade un nuevo juego a una lista existente\n",
                "parameters": [
                    {
                        "name": "gameListName",
                        "in": "path",
                        "description": "Nombre de la lista a obtener información",
                        "required": true,
                        "type": "string",
                        "format": "string"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Juego añadido"
                    },
                    "404": {
                        "description": "Lista no encontrada"
                    },
                    "444": {
                        "description": "Juego ya añadido"
                    }
                }
            }
        },
        "/GameList/{gameListName}/games/{gameName}": {
            "get": {
                "summary": "Juego de una lista",
                "description": "Devuelve el juego especificado de una lista concreta\n",
                "parameters": [
                    {
                        "name": "gameListName",
                        "in": "path",
                        "description": "Nombre de la lista donde esta el juego",
                        "required": true,
                        "type": "string",
                        "format": "string"
                    },
                    {
                        "name": "gameName",
                        "in": "path",
                        "description": "Nombre del juego a obtener informacion",
                        "required": true,
                        "type": "string",
                        "format": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Juego",
                        "schema": {
                            "$ref": "#/definitions/game"
                        }
                    },
                    "404": {
                        "description": "Lista no encontrada"
                    }
                }
            },
            "put": {
                "summary": "Modifica un juego",
                "description": "Modifica un juego concreto de una lista especificada\n",
                "parameters": [
                    {
                        "name": "gameListName",
                        "in": "path",
                        "description": "Nombre de la lista donde esta el juego",
                        "required": true,
                        "type": "string",
                        "format": "string"
                    },
                    {
                        "name": "gameName",
                        "in": "path",
                        "description": "Nombre del juego a modificar",
                        "required": true,
                        "type": "string",
                        "format": "string"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Juego modificado"
                    },
                    "404": {
                        "description": "Lista o juego no encontrado"
                    }
                }
            },
            "delete": {
                "summary": "Elimina un juego",
                "description": "Elimina un juego de una lista especificada\n",
                "parameters": [
                    {
                        "name": "gameListName",
                        "in": "path",
                        "description": "Nombre de la lista donde esta el juego",
                        "required": true,
                        "type": "string",
                        "format": "string"
                    },
                    {
                        "name": "gameName",
                        "in": "path",
                        "description": "Nombre del juego a eliminar",
                        "required": true,
                        "type": "string",
                        "format": "string"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Juego borrado"
                    },
                    "404": {
                        "description": "Lista o juego no encontrado"
                    }
                }
            }
        }
    },
    "definitions": {
        "game": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string",
                    "description": "Nombre del juego en giantBomb."
                },
                "score": {
                    "type": "integer",
                    "description": "Valoración del usuario al juego.ç"
                }
            }
        },
        "gameList": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string",
                    "description": "Nombre de la lista de juegos."
                },
                "description": {
                    "type": "string",
                    "description": "Descripción de una lista de juegos."
                },
                "games": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/game"
                    },
                    "description": "Juegos de la lista"
                }
            }
        }
    }
}